+++ lib/src/office_interops/excel_js_impl.dart
@@ -16,12 +16,17 @@ import 'office_extension_js_impl.dart' as office_extension_js;
 abstract class RequestContextJsImpl
     extends office_core_js.RequestContextJsImpl {
   external WorkbookJsImpl get workbook;
+
+  /// Collection of objects that are tracked for automatic adjustments based
+  /// on surrounding changes in the document.
+  external office_extension_js.TrackedObjectsJsImpl get trackedObjects;
 }
 
 @JS('Workbook')
-abstract class WorkbookJsImpl {
+abstract class WorkbookJsImpl extends office_extension_js.ClientObjectJsImpl {
   /// The request context associated with the object. This connects
   /// the add-in's process to the Office host application's process. */
+  @override
   external RequestContextJsImpl get context;
 
   /// Represents a collection of worksheets associated with the workbook.
@@ -45,9 +50,11 @@ abstract class WorkbookJsImpl {
 }
 
 @JS('WorksheetCollection')
-abstract class WorksheetCollectionJsImpl {
+abstract class WorksheetCollectionJsImpl
+    extends office_extension_js.ClientObjectJsImpl {
   /// The request context associated with the object. This connects
   /// the add-in's process to the Office host application's process. */
+  @override
   external RequestContextJsImpl get context;
 
   /// Gets the loaded child items in this collection. */
@@ -116,7 +123,12 @@ abstract class WorksheetCollectionJsImpl {
   /// To get proper type convert to [WorksheetNameChangedEventArgs]
   external office_extension_js.EventHandlersJsImpl get onNameChanged;
 
-  /// Occurs when a worksheet is moved within a workbook. This event only triggers when a worksheet is directly moved within a workbook. This event doesn't trigger when the position of a worksheet is indirectly changed, such as when a new worksheet is inserted and causes existing worksheets to change positions.
+  /// Occurs when a worksheet is moved within a workbook.
+  /// This event only triggers when a worksheet is directly
+  /// moved within a workbook. This event doesn't trigger when
+  ///  the position of a worksheet is indirectly changed, such
+  /// as when a new worksheet is inserted and causes existing
+  /// worksheets to change positions.
   ///
   /// @remarks
   /// [Api set: ExcelApiOnline 1.1]
@@ -147,9 +159,10 @@ abstract class WorksheetCollectionJsImpl {
 }
 
 @JS('Worksheet')
-abstract class WorksheetJsImpl {
+abstract class WorksheetJsImpl extends office_extension_js.ClientObjectJsImpl {
   /// The request context associated with the object. This connects
   /// the add-in's process to the Office host application's process. */
+  @override
   external RequestContextJsImpl get context;
 
   /// The display name of the worksheet.
@@ -203,4 +216,202 @@ abstract class WorksheetJsImpl {
   ///
   /// Api set: ExcelApi 1.1
   external void activate();
+
+  /// Gets the `Range` object containing the single cell based on row and
+  /// column numbers. The cell can be outside the bounds of its parent range,
+  /// so long as it stays within the worksheet grid.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  ///
+  /// @param row The row number of the cell to be retrieved. Zero-indexed.
+  /// @param column The column number of the cell to be retrieved. Zero-indexed.
+  external RangeJsImpl getCell(final int row, final int column);
+
+  /// Gets the `Range` object beginning at a particular row index and
+  /// column index, and spanning a certain number of rows and columns.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.7]
+  ///
+  /// @param startRow Start row (zero-indexed).
+  /// @param startColumn Start column (zero-indexed).
+  /// @param rowCount Number of rows to include in the range.
+  /// @param columnCount Number of columns to include in the range.
+  external RangeJsImpl getRangeByIndexes(
+    final int startRow,
+    final int startColumn,
+    final int rowCount,
+    final int columnCount,
+  );
+}
+
+/// Range represents a set of one or more contiguous cells such as a cell,
+///  a row, a column, or a block of cells.
+/// To learn more about how ranges are used throughout the API,
+/// start with {@link https://docs.microsoft.com/office/dev/add-ins/excel/excel-add-ins-core-concepts#ranges | Ranges in the Excel JavaScript API}.
+///
+/// @remarks
+/// [Api set: ExcelApi 1.1]
+///
+@JS('Range')
+abstract class RangeJsImpl extends office_extension_js.ClientObjectJsImpl {
+  /// The request context associated with the object. This connects
+  /// the add-in's process to the Office host application's process. */
+  @override
+  external RequestContextJsImpl get context;
+
+  /// Returns the used range of the given range object.
+  /// If there are no used cells within the range, this function
+  /// will throw an `ItemNotFound` error.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  ///
+  /// @param valuesOnly Considers only cells with values as used cells.
+  /// [Api set: ExcelApi 1.2]
+  external RangeJsImpl getUsedRange([final bool? valuesOnly]);
+
+  /// Returns a `Range` object that represents the surrounding region
+  /// for the top-left cell in this range. A surrounding region is
+  /// a range bounded by any combination of blank rows and
+  /// blank columns relative to this range.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.7]
+  external RangeJsImpl getSurroundingRegion();
+
+  /// Gets the `Range` object beginning at a particular row index and
+  /// column index, and spanning a certain number of rows and columns.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.7]
+  ///
+  /// @param startRow Start row (zero-indexed).
+  /// @param startColumn Start column (zero-indexed).
+  /// @param rowCount Number of rows to include in the range.
+  /// @param columnCount Number of columns to include in the range.
+  external RangeJsImpl getRangeByIndexes(
+    final int startRow,
+    final int startColumn,
+    final int rowCount,
+    final int columnCount,
+  );
+
+  /// Gets a row contained in the range.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  ///
+  /// @param row Row number of the range to be retrieved. Zero-indexed.
+  external RangeJsImpl getRow(final int row);
+
+  /// Gets the last cell within the range. For example,
+  /// the last cell of "B2:D5" is "D5".
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external RangeJsImpl getLastCell();
+
+  /// Gets the last column within the range. For example,
+  /// the last column of "B2:D5" is "D2:D5".
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external RangeJsImpl getLastColumn();
+
+  /// Gets the last row within the range. For example,
+  /// the last row of "B2:D5" is "B5:D5".
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external RangeJsImpl getLastRow();
+
+  /// Gets a column contained in the range.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  ///
+  /// @param column Column number of the range to be retrieved. Zero-indexed.
+  external RangeJsImpl getColumn(final int column);
+
+  /// Queues up a command to load the specified properties of the object.
+  /// You must call `context.sync()` before reading the properties.
+  ///
+  /// @param propertyNames A comma-delimited string or an array of strings
+  /// that specify the properties to load.
+  external RangeJsImpl load(final List<String> propertyNames);
+
+  /// Represents the raw values of the specified range.
+  /// The data returned could be a string, number, or boolean.
+  /// Cells that contain an error will return the error string.
+  ///
+  /// If the returned value starts with a plus ("+"), minus ("-"),
+  /// or equal sign ("="), Excel interprets this value as a formula.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external List<dynamic> get values;
+  external set values(final List<dynamic> values);
+
+  /// Returns the total number of rows in the range.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external int get rowCount;
+
+  /// Returns the row number of the first cell in the range. Zero-indexed.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external int get rowIndex;
+
+  /// Specifies the total number of columns in the range.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external int get columnCount;
+
+  /// Specifies the column number of the first cell in the range. Zero-indexed.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external int get columnIndex;
+
+  /// Represents Excel's number format code for the given range.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external List<dynamic> get numberFormat;
+  external set numberFormat(final List<dynamic> value);
+
+  /// Returns a format object, encapsulating the range's font,
+  /// fill, borders, alignment, and other properties.
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external RangeFormatJsImpl get format;
+}
+
+/// A format object encapsulating the range's font, fill, borders,
+/// alignment, and other properties.
+///
+/// @remarks
+/// [Api set: ExcelApi 1.1]
+@JS('RangeFormat')
+abstract class RangeFormatJsImpl
+    extends office_extension_js.ClientObjectJsImpl {
+  /// The request context associated with the object. This connects
+  /// the add-in's process to the Office host application's process. */
+  @override
+  external RequestContextJsImpl get context;
+
+  /// Specifies if Excel wraps the text in the object.
+  /// A `null` value indicates that the entire range
+  /// doesn't have a uniform wrap setting
+  ///
+  /// @remarks
+  /// [Api set: ExcelApi 1.1]
+  external bool get wrapText;
+  external set wrapText(final bool value);
 }