+++ lib/src/office_interops/office_extension_js_impl.dart
@@ -89,3 +89,29 @@ abstract class EventHandlerResultJsImpl {
   external ClientRequestContextJsImpl get context;
   external void remove();
 }
+
+/// Collection of tracked objects, contained within a request context.
+/// See {@link https://learn.microsoft.com/javascript/api/office/officeextension.clientrequestcontext#office-officeextension-clientrequestcontext-trackedobjects-member | context.trackedObjects}
+/// for more information.
+@JS('TrackedObjects')
+abstract class TrackedObjectsJsImpl {
+  /// Track a new object for automatic adjustment based on surrounding
+  /// changes in the document. Only some object types require this.
+  /// If you are using an object across ".sync" calls and outside
+  /// the sequential execution of a ".run" batch,
+  /// and get an "InvalidObjectPath" error when setting a property or
+  /// invoking a method on the object, you needed to have added the object
+  /// to the tracked object collection when the object was first created.
+  /// If this object is part of a collection in Word, you should also track
+  /// the parent collection.
+  external void add(final ClientObjectJsImpl object);
+
+  /// Release the memory associated with an object that was previously
+  /// added to this collection.
+  /// Having many tracked objects slows down the Office application,
+  /// so please remember to free any objects you add, once you're
+  /// done using them.
+  /// You will need to call `context.sync()` before the memory release
+  /// takes effect.
+  external void remove(final ClientObjectJsImpl object);
+}