+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0
+
+- added getCell, getRangeByIndexes to Worksheet
+- added Range class
+- fixed superclasses for Worksheet, Range, Workbook - now they all depends from the ClientObject as it is in Excel API
+- added trackedObjects to RequestContext
+
 ## 0.3.1
 
 - added name property for Excel Workbook